Working on code cleanup ...

+++ bungee/src/main/java/com/dbsoftwares/bungeeutilisals/language/PluginLanguageManager.java
@@ -86,8 +86,7 @@ public class PluginLanguageManager extends AbstractLanguageManager {
             folder.mkdir();
         }
         try {
-            if (!target.exists()) {
-                target.createNewFile();
+            if (!target.exists() && target.createNewFile()) {
                 try (InputStream in = plugin.getResourceAsStream(source); OutputStream out = new FileOutputStream(target)) {
                     if (in == null) {
                         BUCore.getLogger().info("Could not find default language configuration configuration for " +
